Move read access part of IDocumentMetaStoreContext to searchlib.
