+++ searchcore/src/vespa/searchcore/proton/documentmetastore/i_document_meta_store_context.h
@@ -3,29 +3,14 @@
 #pragma once
 
 #include "i_document_meta_store.h"
+#include <vespa/searchlib/common/i_document_meta_store_context.h>
 
 namespace proton {
 
 /**
  * API for providing write and read interface to the document meta store.
  */
-struct IDocumentMetaStoreContext {
-
-    /**
-     * Guard for access to the read interface.
-     * This guard should be alive as long as read interface is used.
-     */
-    struct IReadGuard {
-
-        typedef std::unique_ptr<IReadGuard> UP;
-
-        virtual ~IReadGuard() {}
-
-        /**
-         * Access to read interface.
-         */
-        virtual const search::IDocumentMetaStore &get() const = 0;
-    };
+struct IDocumentMetaStoreContext : public search::IDocumentMetaStoreContext {
 
     typedef std::shared_ptr<IDocumentMetaStoreContext> SP;
 
@@ -35,14 +20,8 @@ struct IDocumentMetaStoreContext {
      * Access to write interface.
      * Should only be used by the writer thread.
      */
-    virtual proton::IDocumentMetaStore & get() = 0;
+    virtual proton::IDocumentMetaStore &get() = 0;
     virtual proton::IDocumentMetaStore::SP getSP() const = 0;
-
-    /**
-     * Access to read interface.
-     * Should be used by all reader threads.
-     */
-    virtual IReadGuard::UP getReadGuard() const = 0;
 
     /**
      * Construct free lists of underlying meta store.
